requirements.txt
# PackNet项目依赖
# 核心框架
jittor>=1.3.0

# 数据处理
numpy>=1.19.0
Pillow>=8.0.0

# 可选依赖
matplotlib>=3.3.0  # 用于可视化
tqdm>=4.60.0       # 进度条
